+++ src/ServerApi/Services/Adnc.Maint/Adnc.Maint.Application/Services/CfgAppService.cs
@@ -69,17 +69,17 @@ namespace Adnc.Maint.Application.Services
         public async Task<PageModelDto<CfgDto>> GetPagedAsync(CfgSearchPagedDto search)
         {
             var whereCondition = ExpressionCreator
-                                                                             .New<CfgDto>()
-                                                                             .AndIf(search.Name.IsNotNullOrWhiteSpace(), x => x.Name.Contains(search.Name))
-                                                                             .AndIf(search.Value.IsNotNullOrWhiteSpace(), x => x.Value.Contains(search.Value));
+                                                .New<CfgDto>()
+                                                .AndIf(search.Name.IsNotNullOrWhiteSpace(), x => x.Name.Contains(search.Name))
+                                                .AndIf(search.Value.IsNotNullOrWhiteSpace(), x => x.Value.Contains(search.Value));
 
             var allCfgs = await _cacheService.GetAllCfgsFromCacheAsync();
 
             var pagedCfgs = allCfgs.Where(whereCondition.Compile())
-                                                   .OrderByDescending(x => x.CreateTime)
-                                                   .Skip((search.PageIndex - 1) * search.PageSize)
-                                                   .Take(search.PageSize)
-                                                   .ToArray();
+                                   .OrderByDescending(x => x.CreateTime)
+                                   .Skip((search.PageIndex - 1) * search.PageSize)
+                                   .Take(search.PageSize)
+                                   .ToArray();
 
             var result = new PageModelDto<CfgDto>()
             {
