+++ src/ServerApi/Services/Adnc.Usr/Adnc.Usr.Application/Services/RoleAppService.cs
@@ -138,8 +138,8 @@ namespace Adnc.Usr.Application.Services
         public async Task<PageModelDto<RoleDto>> GetPagedAsync(RolePagedSearchDto input)
         {
             var whereCondition = ExpressionCreator
-                                                                            .New<SysRole>()
-                                                                            .AndIf(input.RoleName.IsNotNullOrWhiteSpace(), x => x.Name.Contains(input.RoleName));
+                                                  .New<SysRole>()
+                                                  .AndIf(input.RoleName.IsNotNullOrWhiteSpace(), x => x.Name.Contains(input.RoleName));
 
             var pagedModel = await _roleRepository.PagedAsync(input.PageIndex, input.PageSize, whereCondition, x => x.Ordinal, true);
