+++ src/ServerApi/Services/Shared/Adnc.Application.Shared/DtoValidator/PhonePropertyValidator.cs
@@ -7,6 +7,8 @@ namespace Adnc.Application.Shared.DtoValidators
 {
     public class PhonePropertyValidator<T> : PropertyValidator<T, string>
     {
+        public static string Pattern => @"^((1[3,5,6,8][0-9])|(14[5,7])|(17[0,1,3,6,7,8])|(19[8,9]))\d{8}$";
+        public static string ErrorMessage => @"不是有效的手机号码。";
         public override string Name => nameof(PhonePropertyValidator<T>);
 
         public override bool IsValid(ValidationContext<T> context, string value)
@@ -14,13 +16,13 @@ namespace Adnc.Application.Shared.DtoValidators
             if (value.IsNullOrWhiteSpace())
                 return false;
 
-            Match match = Regex.Match(value, @"^((1[3,5,6,8][0-9])|(14[5,7])|(17[0,1,3,6,7,8])|(19[8,9]))\d{8}$");
+            var match = Regex.Match(value, Pattern);
 
             return match.Success;
         }
 
         protected override string GetDefaultMessageTemplate(string errorCode)
-            => "'{PropertyName}' 不是有效的手机号码." + errorCode;
+            => "'{PropertyName}' " + ErrorMessage + " " + errorCode;
     }
 
     public static class PhoneValidatorExtensions
